gestion des erreurs quand on reference un post inexistant

+++ src/Casscroute/BlogBundle/Controller/CrudController.php
@@ -41,6 +41,9 @@ class CrudController extends Controller
         $em = $this->getDoctrine()->getManager();
         $postRepository = $em->getRepository('CasscrouteBlogBundle:Post');
         $post = $postRepository->findOneByUrlAlias($alias);
+        if (!$post) {
+            throw $this->createNotFoundException('Le Post n\'existe pas.');
+        }
 
         if ($post == null) {
             $request->getSession()->getFlashBag()->add('error', "Ce post n'existe pas.");
@@ -71,6 +74,9 @@ class CrudController extends Controller
         $em = $this->getDoctrine()->getManager();
         $postRepository = $em->getRepository('CasscrouteBlogBundle:Post');
         $post = $postRepository->findOneByUrlAlias($alias);
+        if (!$post) {
+            throw $this->createNotFoundException('Le Post n\'existe pas.');
+        }
 
         if ($post != null) {
             $em->remove($post);
